fix: wrap scan and campaign pages content with suspense for not found
