+++ src/app/[lng]/campaign/[slug]/page.tsx
@@ -1,13 +1,11 @@
+import { getCampaigns } from '@/app/lib/getCampaigns';
 import type { Metadata } from 'next';
-import { notFound } from 'next/navigation';
 import { Suspense } from 'react';
-import { getCampaigns } from 'src/app/lib/getCampaigns';
 import { getCampaignBySlug } from 'src/app/lib/getCampaignsBySlug';
 import { siteName } from 'src/app/lib/metadata';
-import { CampaignPage } from 'src/components/Campaign/CampaignPage';
+import { CampaignPage } from '@/components/Campaign/CampaignPage';
 import { CampaignPageSkeleton } from 'src/components/Campaign/CampaignPageSkeleton';
 import { getSiteUrl } from 'src/const/urls';
-import { fetchQuestOpportunitiesByRewardsIds } from 'src/utils/merkl/fetchQuestOpportunities';
 import { sliceStrToXChar } from 'src/utils/splitStringToXChar';
 
 // Add generateStaticParams function
@@ -31,7 +29,7 @@ export async function generateMetadata({
   params,
 }: {
   params: Promise<{ slug: string }>;
-}): Promise<Metadata> {
+}) {
   const { slug } = await params;
 
   try {
@@ -57,11 +55,14 @@ export async function generateMetadata({
       alternates: {
         canonical: `${getSiteUrl()}/campaign/${slug}`,
       },
+      other: {
+        'partner-theme': slug,
+      },
       twitter: openGraph,
       openGraph,
     };
   } catch (err) {
-    notFound();
+    return {};
   }
 }
 
@@ -69,19 +70,10 @@ type Params = Promise<{ slug: string }>;
 
 export default async function Page({ params }: { params: Params }) {
   const { slug } = await params;
-  const campaign = await getCampaignBySlug(slug);
-
-  if (!campaign || !campaign.data || campaign.data.length === 0) {
-    notFound();
-  }
-
-  const extendedQuests = await fetchQuestOpportunitiesByRewardsIds(
-    campaign.data[0].quests,
-  );
 
   return (
     <Suspense fallback={<CampaignPageSkeleton />}>
-      <CampaignPage campaign={campaign.data[0]} quests={extendedQuests} />
+      <CampaignPage slug={slug} />
     </Suspense>
   );
 }
